Added more samples to sample.txt and implemented basic dump load metric

+++ data.py
@@ -70,7 +70,7 @@ class Data(object):
             f.readline()
 
         # Very simple placeholder for timeList. Will have to be modified later.
-        self.timeList = np.linspace(0, self.samplingPeriod, self.samplingPeriod / self.samplingRate)
+        self.timeList = np.linspace(0, self.samplingPeriod, int(self.samplingPeriod / self.samplingRate))
 
         f.close()
